move slack slash command to app module
