+++ busy_beaver/blueprints/slack/toolbox.py
@@ -1,14 +1,12 @@
-from flask import jsonify
+"""Slack toolbox"""
 
 
 def make_slack_response(
     response_type="ephemeral", text="", attachments=None, blocks=None
 ):
-    return jsonify(
-        {
-            "response_type": response_type,
-            "text": text,
-            "attachments": [attachments] if attachments else [],
-            "blocks": blocks if blocks else [],
-        }
-    )
+    return {
+        "response_type": response_type,
+        "text": text,
+        "attachments": [attachments] if attachments else [],
+        "blocks": blocks if blocks else [],
+    }
